Changes from library code
git-svn-id: svn://localhost/vantis/repgen/trunk@351 f059b2e8-4df0-46e1-8ac8-86f86453394b

+++ vsprojects/repgen/App_Code/DataLayer/Model.cs
@@ -30,7 +30,7 @@ namespace RSMTenon.Data
                             select new AssetWeighting
                             {
                                 AssetClass = g.Key,
-                                Weighting = g.Sum(model => model.Weighting)
+                                Weighting = (float)g.Sum(model => model.Weighting)
                             };
 
             return alloction;
